requirements.txt
tensorflow
tensorflow-datasets
streamlit
numpy
pillow
matplotlib